+++ src/main/java/fiuba/tpp/reactorapp/model/response/chemicalmodels/ModelResponse.java
@@ -1,0 +1,34 @@
+package fiuba.tpp.reactorapp.model.response.chemicalmodels;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import fiuba.tpp.reactorapp.model.math.Observation;
+
+import java.util.List;
+
+public class ModelResponse {
+
+    @JsonProperty("r2")
+    private double rms;
+
+    @JsonProperty("observaciones")
+    private List<Observation> observations;
+
+    public ModelResponse() {
+    }
+
+    public double getRms() {
+        return rms;
+    }
+
+    public void setRms(double rms) {
+        this.rms = rms;
+    }
+
+    public List<Observation> getObservations() {
+        return observations;
+    }
+
+    public void setObservations(List<Observation> observations) {
+        this.observations = observations;
+    }
+}
